Feat: hashtags shares
